+++ frontend/src/metabase-types/api/collection.ts
@@ -1,8 +1,3 @@
 export interface Collection {
   id: number;
 }
-
-export const createCollection = (opts?: Partial<Collection>): Collection => ({
-  id: 1,
-  ...opts,
-});
