+++ frontend/src/metabase-types/api/database.ts
@@ -8,13 +8,3 @@ export interface Database {
   creator_id?: number;
   initial_sync_status: InitialSyncStatus;
 }
-
-export const createDatabase = (opts?: Partial<Database>): Database => ({
-  id: 1,
-  name: "Database",
-  engine: "H2",
-  is_sample: false,
-  creator_id: undefined,
-  initial_sync_status: "complete",
-  ...opts,
-});
