+++ frontend/src/metabase-types/api/mocks/automagic-dashboards.ts
@@ -1,0 +1,18 @@
+import { DatabaseCandidate, TableCandidate } from "metabase-types/api";
+
+export const createDatabaseCandidate = (
+  opts?: Partial<DatabaseCandidate>,
+): DatabaseCandidate => ({
+  id: "1/public",
+  schema: "public",
+  tables: [],
+  ...opts,
+});
+
+export const createTableCandidate = (
+  opts?: Partial<TableCandidate>,
+): TableCandidate => ({
+  title: "Sample table",
+  url: "/auto/1",
+  ...opts,
+});
