+++ frontend/src/metabase-types/api/mocks/collection.ts
@@ -1,0 +1,6 @@
+import { Collection } from "metabase-types/api";
+
+export const createCollection = (opts?: Partial<Collection>): Collection => ({
+  id: 1,
+  ...opts,
+});
