+++ frontend/src/metabase-types/api/mocks/dashboard.ts
@@ -1,0 +1,7 @@
+import { Dashboard } from "metabase-types/api";
+
+export const createDashboard = (opts?: Partial<Dashboard>): Dashboard => ({
+  id: 1,
+  name: "Dashboard",
+  ...opts,
+});
