+++ frontend/src/metabase-types/api/mocks/index.ts
@@ -1,0 +1,5 @@
+export * from "./automagic-dashboards";
+export * from "./collection";
+export * from "./dashboard";
+export * from "./database";
+export * from "./user";
