+++ frontend/src/metabase-types/api/user.ts
@@ -5,12 +5,3 @@ export interface User {
   has_invited_second_user: boolean;
   personal_collection_id: number;
 }
-
-export const createUser = (opts?: Partial<User>): User => ({
-  id: 1,
-  first_name: "Testy",
-  is_superuser: false,
-  has_invited_second_user: false,
-  personal_collection_id: 1,
-  ...opts,
-});
