+++ frontend/src/metabase-types/store/mocks/settings.ts
@@ -1,0 +1,8 @@
+import { SettingsState } from "metabase-types/store";
+
+export const createSettingsState = (
+  opts?: Partial<SettingsState>,
+): SettingsState => ({
+  values: {},
+  ...opts,
+});
