+++ frontend/src/metabase-types/store/mocks/state.ts
@@ -1,0 +1,7 @@
+import { State } from "metabase-types/store";
+import { createSettingsState } from "metabase-types/store/mocks";
+
+export const createState = (opts?: Partial<State>): State => ({
+  settings: createSettingsState(),
+  ...opts,
+});
