+++ frontend/src/metabase-types/store/settings.ts
@@ -1,10 +1,3 @@
 export interface SettingsState {
   values: Record<string, unknown>;
 }
-
-export const createSettingsState = (
-  opts?: Partial<SettingsState>,
-): SettingsState => ({
-  values: {},
-  ...opts,
-});
