+++ frontend/src/metabase-types/store/state.ts
@@ -1,10 +1,5 @@
-import { createSettingsState, SettingsState } from "./settings";
+import { SettingsState } from "./settings";
 
 export interface State {
   settings: SettingsState;
 }
-
-export const createState = (opts?: Partial<State>): State => ({
-  settings: createSettingsState(),
-  ...opts,
-});
