+++ frontend/src/metabase/home/homepage/components/XraySection/XraySection.unit.spec.tsx
@@ -5,7 +5,7 @@ import {
   createDatabaseCandidate,
   createTableCandidate,
   createUser,
-} from "metabase-types/api";
+} from "metabase-types/api/mocks";
 import XraySection from "./XraySection";
 
 describe("XraySection", () => {
