+++ frontend/src/metabase/status/components/DatabaseStatus/DatabaseStatus.unit.spec.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { render, screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
-import { createDatabase, createUser } from "metabase-types/api";
+import { createDatabase, createUser } from "metabase-types/api/mocks";
 import DatabaseStatus from "./DatabaseStatus";
 
 describe("DatabaseStatus", () => {
